third_party/WebKit/LayoutTests/http/tests/permissions/iframe-geolocation-permissions.html

<!DOCTYPE html>
<html>
<head>
<title>Cross-Origin Geolocation With Improper Permissions Attribute</title>
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}
internals.forceRestrictIFramePermissions();
</script>
</head>
<body>
<iframe src="resources/geolocation-user.html" permissions=""></iframe>
</body>
</html>
